refactor: move LibraryNamespace into library module
