+++ assembly/src/library/masl.rs
@@ -1,49 +1,8 @@
 use super::{
-    ByteReader, ByteWriter, Deserializable, DeserializationError, LibraryError, LibraryNamespace,
-    LibraryPath, ModuleAst, Serializable, Vec,
+    ByteReader, ByteWriter, Deserializable, DeserializationError, Library, LibraryError,
+    LibraryNamespace, LibraryPath, Module, ModuleAst, Serializable, Vec, Version,
 };
-use core::{cmp::Ordering, fmt, slice::Iter};
-
-// LIBRARY
-// ================================================================================================
-
-/// A library definition that provides AST modules for the compilation process.
-pub trait Library {
-    /// The concrete type used to iterate the modules of the library.
-    type ModuleIterator<'a>: Iterator<Item = &'a Module>
-    where
-        Self: 'a;
-
-    /// Returns the root namespace of this library.
-    fn root_ns(&self) -> &LibraryNamespace;
-
-    /// Returns the version number of this library.
-    fn version(&self) -> &Version;
-
-    /// Iterate the modules available in the library.
-    fn modules(&self) -> Self::ModuleIterator<'_>;
-}
-
-impl<T> Library for &T
-where
-    T: Library,
-{
-    type ModuleIterator<'a> = T::ModuleIterator<'a>
-    where
-        Self: 'a;
-
-    fn root_ns(&self) -> &LibraryNamespace {
-        T::root_ns(self)
-    }
-
-    fn version(&self) -> &Version {
-        T::version(self)
-    }
-
-    fn modules(&self) -> Self::ModuleIterator<'_> {
-        T::modules(self)
-    }
-}
+use core::slice::Iter;
 
 // LIBRARY IMPLEMENTATION FOR MASL FILES
 // ================================================================================================
@@ -268,217 +227,5 @@ impl Deserializable for MaslLibrary {
             version,
             modules,
         })
-    }
-}
-
-// MODULE
-// ================================================================================================
-
-/// A module containing its absolute path and parsed AST.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Module {
-    /// Absolute path of the module.
-    pub path: LibraryPath,
-    /// Parsed AST of the module.
-    pub ast: ModuleAst,
-}
-
-impl Module {
-    // CONSTRUCTORS
-    // --------------------------------------------------------------------------------------------
-
-    /// Create a new module from a path and ast.
-    pub const fn new(path: LibraryPath, ast: ModuleAst) -> Self {
-        Self { path, ast }
-    }
-
-    /// Create a new kernel module from a AST using the constant [`LibraryPath::kernel_path`].
-    pub fn kernel(ast: ModuleAst) -> Self {
-        Self {
-            path: LibraryPath::kernel_path(),
-            ast,
-        }
-    }
-
-    // VALIDATIONS
-    // --------------------------------------------------------------------------------------------
-
-    /// Validate if the module belongs to the provided namespace.
-    pub fn check_namespace(&self, namespace: &LibraryNamespace) -> Result<(), LibraryError> {
-        (self.path.first() == namespace.as_str())
-            .then_some(())
-            .ok_or_else(|| LibraryError::namespace_violation(self.path.first(), namespace.as_str()))
-    }
-}
-
-impl PartialOrd for Module {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.path.partial_cmp(&other.path)
-    }
-}
-
-impl Ord for Module {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.path.cmp(&other.path)
-    }
-}
-
-impl Serializable for Module {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        self.path.write_into(target);
-        self.ast.write_into(target);
-    }
-}
-
-impl Deserializable for Module {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let path = LibraryPath::read_from(source)?;
-        let ast = ModuleAst::read_from(source)?;
-        Ok(Self { path, ast })
-    }
-}
-
-// VERSION
-// ================================================================================================
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct Version {
-    pub major: u16,
-    pub minor: u16,
-    pub patch: u16,
-}
-
-impl Version {
-    // CONSTANTS
-    // --------------------------------------------------------------------------------------------
-
-    /// Minimal version
-    pub const MIN: Self = Self {
-        major: 0,
-        minor: 1,
-        patch: 0,
-    };
-
-    // COMPARISON HELPERS
-    // --------------------------------------------------------------------------------------------
-
-    /// Compare two versions considering only the major value.
-    pub fn cmp_major(&self, other: &Self) -> Ordering {
-        self.major.cmp(&other.major)
-    }
-
-    /// Compare two versions considering only the major and minor values.
-    pub fn cmp_minor(&self, other: &Self) -> Ordering {
-        match self.cmp_major(other) {
-            Ordering::Less => Ordering::Less,
-            Ordering::Equal => self.minor.cmp(&other.minor),
-            Ordering::Greater => Ordering::Greater,
-        }
-    }
-
-    /// Compare two versions considering the major, minor, and patch values.
-    pub fn cmp_patch(&self, other: &Self) -> Ordering {
-        match self.cmp_minor(other) {
-            Ordering::Less => Ordering::Less,
-            Ordering::Equal => self.patch.cmp(&other.patch),
-            Ordering::Greater => Ordering::Greater,
-        }
-    }
-
-    // INCREMENT HELPERS
-    // --------------------------------------------------------------------------------------------
-
-    /// Increment the major version value.
-    pub const fn inc_major(self) -> Self {
-        Self {
-            major: self.major + 1,
-            minor: 0,
-            patch: 0,
-        }
-    }
-
-    /// Increment the minor version value.
-    pub const fn inc_minor(self) -> Self {
-        Self {
-            major: self.major,
-            minor: self.minor + 1,
-            patch: 0,
-        }
-    }
-
-    /// Increment the patch version value.
-    pub const fn inc_patch(self) -> Self {
-        Self {
-            major: self.major,
-            minor: self.minor,
-            patch: self.patch + 1,
-        }
-    }
-}
-
-impl Default for Version {
-    fn default() -> Self {
-        Self::MIN
-    }
-}
-
-impl Serializable for Version {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        target.write_u16(self.major);
-        target.write_u16(self.minor);
-        target.write_u16(self.patch);
-    }
-}
-
-impl Deserializable for Version {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let major = source.read_u16()?;
-        let minor = source.read_u16()?;
-        let patch = source.read_u16()?;
-        Ok(Self {
-            major,
-            minor,
-            patch,
-        })
-    }
-}
-
-impl fmt::Display for Version {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}.{}.{}", self.major, self.minor, self.patch)
-    }
-}
-
-impl TryFrom<&str> for Version {
-    type Error = LibraryError;
-
-    fn try_from(value: &str) -> Result<Self, Self::Error> {
-        let mut components = value.split('.');
-
-        let major = components
-            .next()
-            .ok_or(LibraryError::missing_version_component(value, "major"))?
-            .parse::<u16>()
-            .map_err(|err| LibraryError::invalid_version_number(value, err.to_string()))?;
-        let minor = components
-            .next()
-            .ok_or(LibraryError::missing_version_component(value, "minor"))?
-            .parse::<u16>()
-            .map_err(|err| LibraryError::invalid_version_number(value, err.to_string()))?;
-        let patch = components
-            .next()
-            .ok_or(LibraryError::missing_version_component(value, "patch"))?
-            .parse::<u16>()
-            .map_err(|err| LibraryError::invalid_version_number(value, err.to_string()))?;
-
-        if components.next().is_some() {
-            Err(LibraryError::too_many_version_components(value))
-        } else {
-            Ok(Self {
-                major,
-                minor,
-                patch,
-            })
-        }
     }
 }
